fix: eslint fast refresh warning
